+++ omniisaacgymenvs/robots/articulations/carterv2.py
@@ -1,0 +1,116 @@
+# Copyright (c) 2018-2022, NVIDIA Corporation
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+#
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+
+from typing import Optional, Sequence
+
+import carb
+import numpy as np
+import torch
+
+from omni.isaac.core.utils.nucleus import get_assets_root_path
+from omni.isaac.core.utils.stage import add_reference_to_stage
+from omni.isaac.core.prims.rigid_prim import RigidPrim
+from omni.isaac.core.utils.extensions import enable_extension
+enable_extension("omni.isaac.wheeled_robots")
+from omni.isaac.wheeled_robots.controllers.differential_controller import DifferentialController
+from omni.isaac.wheeled_robots.robots import WheeledRobot
+from omni.isaac.core.utils.stage import get_current_stage
+
+class CarterV2(WheeledRobot):
+    def __init__(
+        self,
+        prim_path: str,
+        name: Optional[str] = "CarterV2",
+        wheel_dof_names=["joint_wheel_left", "joint_wheel_right"],
+        create_robot=True,
+        usd_path: Optional[str] = None,
+        position: Optional[np.ndarray] = None,
+        orientation: Optional[np.ndarray] = None,
+    ) -> None:
+
+        self._usd_path = usd_path
+        self._name = name
+
+        if self._usd_path is None:
+            assets_root_path = get_assets_root_path()
+            if assets_root_path is None:
+                carb.log_error("Could not find Isaac Sim assets folder")
+            self._usd_path = assets_root_path + "/Isaac/Robots/Carter/nova_carter.usd" # # Nova carter usd
+
+        add_reference_to_stage(self._usd_path, prim_path)
+
+        super().__init__(
+            prim_path=prim_path,
+            name=name,
+            wheel_dof_names=wheel_dof_names,
+            create_robot=create_robot,
+            position=position,
+            orientation=orientation,
+        )
+        ## parameters read from carterv2.urdf
+        self.controller = DifferentialController(name="simple_control", wheel_radius=0.135, wheel_base=0.336)
+
+class NvidiaCube(RigidPrim):
+    def __init__(
+        self,
+        prim_path: str,
+        name: Optional[str] = "NvidiaCube",
+        usd_path: Optional[str] = None,
+        translation: Optional[np.ndarray] = None,
+        orientation: Optional[np.ndarray] = None,
+        scale: Optional[Sequence[float]] = None,
+    )-> None:
+        
+        self._usd_path = usd_path
+        self._name = name
+
+        if self._usd_path is None:
+            assets_root_path = get_assets_root_path()
+            if assets_root_path is None:
+                carb.log_error("Could not find Isaac Sim assets folder")
+            self._usd_path = assets_root_path + "/Isaac/Props/Blocks/nvidia_cube.usd" 
+
+        add_reference_to_stage(self._usd_path, prim_path)
+
+        
+        from omni.isaac.core.utils.prims import define_prim
+        from omni.isaac.core.utils.semantics import add_update_semantics
+        ## Get target prim in stage
+        stage = get_current_stage()
+        prim = stage.GetPrimAtPath(prim_path)
+        ## Add semantic label
+        add_update_semantics(prim, semantic_label=name, type_label="class")
+
+        super().__init__(
+            prim_path=prim_path,
+            name=name,
+            translation=translation,
+            orientation=orientation,
+            scale=scale,
+        )